Move BRLA onramp hold helper out of handlers
